Remove API routes
