+++ app/page.tsx
@@ -3,21 +3,18 @@ import VerticalProductCarousel from "@/components/VerticalProductCarousel";
 import { Product } from "@/models/Product";
 import { Suspense } from "react";
 import Loading from "./loading";
+import { promises as fs } from "fs";
 
 const getProductsList = async () => {
-  const BASE_URL = process.env.VERCEL_ENV
-    ? process.env.VERCEL_URL
-    : process.env.URL;
-  const productsListRes = await fetch(`${BASE_URL}/api/products`);
-  if (!productsListRes.ok) {
-    throw new Error("Failed to fetch products list");
-  }
-
-  return productsListRes.json();
+  const fileData = await fs.readFile(
+    process.cwd() + "/mocks/products-list.json",
+    "utf8"
+  );
+  return JSON.parse(fileData);
 };
 
 const Home = async () => {
-  const productsList: Product[] = JSON.parse(await getProductsList());
+  const productsList: Product[] = await getProductsList();
 
   return (
     <Suspense fallback={<Loading />}>
